feat: Single ViewModel for Login and Registration flows
