+++ app/src/main/kotlin/com/blox/payments/ui/registration/RegistrationPhoneNumberScreen.kt
@@ -1,8 +1,9 @@
-package com.blox.payments.ui.registration.password
+package com.blox.payments.ui.registration
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
@@ -10,25 +11,25 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardCapitalization
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.hilt.navigation.compose.hiltViewModel
 import com.blox.payments.R
+import com.blox.payments.domain.registration.model.Countries
 import com.blox.uicomponents.commons.FullWidthButton
-import com.blox.uicomponents.commons.FullWidthPasswordTextField
+import com.blox.uicomponents.commons.FullWidthTextField
 import com.blox.uicomponents.commons.ScreenTitle
 import kotlinx.coroutines.launch
 
 @Composable
-fun RegistrationPasswordScreen(
-    onSuccessfulEmail: () -> Unit
+fun RegistrationPhoneNumberScreen(
+    viewModel: RegistrationViewModel,
+    onSuccessfulPhoneNumber: () -> Unit
 ) {
-    val viewModel = hiltViewModel<RegistrationPasswordViewModel>()
     val uiState = viewModel.uiState
 
     LaunchedEffect(uiState) {
         launch {
-            if (uiState.passwordCompleted) {
-                viewModel.passwordCompletedHandled()
-                onSuccessfulEmail()
+            if (uiState.phoneNumberCompleted) {
+                viewModel.phoneNumberCompletedHandled()
+                onSuccessfulPhoneNumber()
             }
         }
     }
@@ -36,20 +37,25 @@ fun RegistrationPasswordScreen(
     Column(
         modifier = Modifier.fillMaxSize()
     ) {
-        ScreenTitle(title = stringResource(id = R.string.registration_password_title))
-        FullWidthPasswordTextField(
-            value = viewModel.password,
-            hint = stringResource(id = R.string.registration_password_label),
+        ScreenTitle(title = stringResource(id = R.string.registration_phone_number_title))
+        FullWidthTextField(
+            value = viewModel.phoneNumber,
+            hint = stringResource(id = R.string.registration_phone_number_label),
+            prefix = {
+                Text(
+                    Countries.getCountryCodeByName(viewModel.country)
+                )
+            },
             keyboardOptions = KeyboardOptions(
                 capitalization = KeyboardCapitalization.None,
-                keyboardType = KeyboardType.Password,
+                keyboardType = KeyboardType.Phone,
                 imeAction = ImeAction.Done
             )
         ) {
-            viewModel.updatePassword(it)
+            viewModel.phoneNumber = it
         }
         FullWidthButton(text = stringResource(id = R.string.continue_cta)) {
-            viewModel.validatePassword()
+            viewModel.validatePhoneNumber()
         }
     }
 }
