+++ app/src/main/kotlin/com/blox/payments/ui/registration/RegistrationUiState.kt
@@ -5,5 +5,13 @@ data class RegistrationUiState(
     val firstNameError: Boolean = false,
     val lastNameError: Boolean = false,
     val birthDateCompleted: Boolean = false,
-    val birthDateError: Boolean = false
+    val birthDateError: Boolean = false,
+    val countryCompleted: Boolean = false,
+    val countryError: Boolean = false,
+    val phoneNumberCompleted: Boolean = false,
+    val phoneNumberError: Boolean = false,
+    val emailCompleted: Boolean = false,
+    val emailError: Boolean = false,
+    val passwordCompleted: Boolean = false,
+    val passwordError: Boolean = false
 )
