--- app/src/main/kotlin/com/blox/payments/ui/registration/country/RegistrationCountryUiState.kt
@@ -1,6 +1,0 @@
-package com.blox.payments.ui.registration.country
-
-data class RegistrationCountryUiState(
-    val countryCompleted: Boolean = false,
-    val countryError: Boolean = false
-)
