--- app/src/main/kotlin/com/blox/payments/ui/registration/email/RegistrationEmailState.kt
@@ -1,6 +1,0 @@
-package com.blox.payments.ui.registration.email
-
-data class RegistrationEmailState(
-    val emailCompleted: Boolean = false,
-    val emailError: Boolean = false
-)
