--- app/src/main/kotlin/com/blox/payments/ui/registration/password/RegistrationPasswordState.kt
@@ -1,6 +1,0 @@
-package com.blox.payments.ui.registration.password
-
-data class RegistrationPasswordState(
-    val passwordCompleted: Boolean = false,
-    val passwordError: Boolean = false
-)
