--- app/src/main/kotlin/com/blox/payments/ui/registration/phonenumber/RegistrationPhoneNumberState.kt
@@ -1,6 +1,0 @@
-package com.blox.payments.ui.registration.phonenumber
-
-data class RegistrationPhoneNumberState(
-    val phoneNumberCompleted: Boolean = false,
-    val phoneNumberError: Boolean = false
-)
